story.txt
Spring break, a time to travel to warm places and enjoy the beach. 
There are always lots of fun memories that come from this time of year. 
This isn't the case every time. It was 1986 and we traveled down to Palm Springs Florida. 
It was a bright and sunny day and I was ready to hit the beach, until my plans were flipped upside down.